+++ src/main/java/com/meoguri/linkocean/domain/user/entity/vo/OAuthType.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.user.entity;
+package com.meoguri.linkocean.domain.user.entity.vo;
 
 /**
  * OAuth 지원 벤더 종류
