+++ src/test/java/com/meoguri/linkocean/domain/user/entity/UserTest.java
@@ -4,6 +4,9 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
+import com.meoguri.linkocean.domain.user.entity.vo.Email;
+import com.meoguri.linkocean.domain.user.entity.vo.OAuthType;
+
 class UserTest {
 
 	@Test
